+++ src/Data/Morpheus/Server/Internal/TH/Decode.hs
@@ -6,13 +6,11 @@
 
 module Data.Morpheus.Server.Internal.TH.Decode
   ( withInputObject,
-    withMaybe,
-    withList,
-    withRefinedList,
     withEnum,
     withInputUnion,
     decodeFieldWith,
     withScalar,
+    handleEither,
   )
 where
 
@@ -51,32 +49,7 @@ withInputObject ::
 withInputObject f (Object object) = f object
 withInputObject _ isType = failure (typeMismatch "InputObject" isType)
 
-withMaybe :: Monad m => (ValidValue -> m a) -> ValidValue -> m (Maybe a)
-withMaybe _ Null = pure Nothing
-withMaybe decode x = Just <$> decode x
-
-withList ::
-  (Failure InternalError m, Monad m) =>
-  (ValidValue -> m a) ->
-  ValidValue ->
-  m [a]
-withList decode (List li) = traverse decode li
-withList _ isType = failure (typeMismatch "List" isType)
-
 -- | Useful for more restrictive instances of lists (non empty, size indexed etc)
-withRefinedList ::
-  (Failure InternalError m, Monad m) =>
-  ([a] -> Either Message (rList a)) ->
-  (ValidValue -> m a) ->
-  ValidValue ->
-  m (rList a)
-withRefinedList refiner decode (List li) = do
-  listRes <- traverse decode li
-  case refiner listRes of
-    Left err -> failure (typeMismatch err (List li))
-    Right value -> pure value
-withRefinedList _ _ isType = failure (typeMismatch "List" isType)
-
 withEnum :: Failure InternalError m => (TypeName -> m a) -> Value VALID -> m a
 withEnum decode (Enum value) = decode value
 withEnum _ isType = failure (typeMismatch "Enum" isType)
@@ -98,7 +71,7 @@ withScalar ::
   (ScalarValue -> Either Token a) ->
   Value VALID ->
   m a
-withScalar typename parseValue value = case toScalar value >>= parseValue of
+withScalar typename decodeScalar value = case toScalar value >>= decodeScalar of
   Right scalar -> pure scalar
   Left message ->
     failure
@@ -109,6 +82,9 @@ withScalar typename parseValue value = case toScalar value >>= parseValue of
 
 decodeFieldWith :: (Value VALID -> m a) -> FieldName -> ValidObject -> m a
 decodeFieldWith decoder = selectOr (decoder Null) (decoder . entryValue)
+
+handleEither :: Failure InternalError m => Either Message a -> m a
+handleEither = either (failure . msgInternal) pure
 
 -- if value is already validated but value has different type
 typeMismatch :: Message -> Value s -> InternalError